PRC-219: Create contact API (#75)
Also removed the primary_phone field as it could not be set but was being used in the contact summary view which meant you would never get a phone number using real data. Changed to a window function to get last created.

Also, also, added the phone fields to the contact prisoner list as they were being queried and not returned or tested.