PRC-74: Rename restriction URLs (#166)
Updated urls to singular for as per other entities (restriction instead of restrictions) and also removed estate-wide- from contact restrictions. as the URL is already under /contact/:id/ it is clear that they relate only to the contact.

Moving away from use of estate-wide as they are referred to as global in the UI. This is subject to change again but updated the docs to reference both "global" and estate-wide.